requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
rich>=13.0.0
openai>=1.0.0
anthropic>=0.40.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
